break up the huge fsm.go file, introduce the FSMSerializer interface to break the circular dep between FSM and FSMContext
